update solution with double

+++ LoadDataIntoTableViewPractice/FibonacciNumber.swift
@@ -1,9 +1,9 @@
 import Foundation
 
 struct FibonacciNumber {
-    static private var cache = [Int: Int]()
+    static private var cache = [Int: Double]()
   
-    static func fib(n: Int) -> Int {
+    static func fib(n: Int) -> Double {
         if let val = cache[n] { return val }
         if n == 0 || n == 1 { return 1 }
         let val = fib(n: n - 1) + fib(n: n - 2)
